Move SessionStore to PayPal module.
